Use BrickLink IDs as search queries
Rebrickable tells us the BrickLink ID of a part, but doesn't say if
BrickLink considers it a part or a minifig. The URLs for the two are
different, so and searching `P=foo` won't work if it's a minifig.

So instead, always go to the search page. It's not as nice as opening
the details page directly, but it's much more reliable.