Ensure the thumbnail box always has an image
`url` is a required field for the Discord Thumbnail component. If
Rebrickable didn't give us an image URL, and we blindly pass
`undefined` here, then Discord complains and we crash.

The crashing is because we don't handle exceptions in async functions
properly - that's a separate fix that's very necessary. But this solves
the immediate issue.