Gracefully handle DiscordMessage errors
Because DiscordRequest is asynchronous, we wouldn't catch any errors in
our outer `try {}`, meaning we would crash. Oops!

So, catch the errors properly, and then both log them (with hopefully
enough information to be able to debug the issue) and send a message to
Discord so that people know what happened.