Restructure commands as classes
We now have a Command base class. Commands are responsible for providing
the definition to Discord, and have run() and autocomplete() methods
for executing the command and providing autocomplete data.

The code in app.js is now a lot simpler, and only worries about
initialization, and then looking up and calling the relevant Command.