Reload SSL certificates when they change
When we start up, we load the SSL certificate. However, that certificate
expires after a while, and so we need to load the fresh one whenever
it's renewed. This code should do that.

This is made a little more complicated by the nuances of fs.watch() and
symlinks: We resolve any symlinks in the key path and then watch its
parent directory, to make sure that we get the change events.

I'm not 100% sure if this will solve the regular downtime issues,
because the other half of that is whether the certificate is actually
getting renewed or not. But the logging here will let me see that at
least.